+++ index.js
@@ -43,20 +43,52 @@ function parseUpdateArguments (conditions, doc, options, callback) {
     return args;
 }
 
+function parseIndexFields (options) {
+    var indexFields = {
+        deleted: false,
+        deletedAt: false,
+        deletedBy: false
+    };
+
+    if (!options.indexFields) {
+        return indexFields;
+    }
+
+    if ((typeof options.indexFields === 'string' || options.indexFields instanceof String) && options.indexFields === 'all') {
+        indexFields.deleted = indexFields.deletedAt = indexFields.deletedBy = true;
+    }
+
+    if (typeof(options.indexFields) === "boolean" && options.indexFields === true) {
+        indexFields.deleted = indexFields.deletedAt = indexFields.deletedBy = true;
+    }
+
+    if (Array.isArray(options.indexFields)) {
+        indexFields.deleted = options.indexFields.indexOf('deleted') > -1;
+        indexFields.deletedAt = options.indexFields.indexOf('deletedAt') > -1;
+        indexFields.deletedBy = options.indexFields.indexOf('deletedBy') > -1;
+    }
+
+    return indexFields;
+}
+
 module.exports = function (schema, options) {
+    options = options || {};
+    var indexFields = parseIndexFields(options)
+
     schema.add({
         deleted: {
             type: Boolean,
-            default: false
+            default: false,
+            index: indexFields.deleted
         }
     });
 
-    if (options && options.deletedAt === true) {
-        schema.add({ deletedAt: { type: Date} });
+    if (options.deletedAt === true) {
+        schema.add({ deletedAt: { type: Date, index: indexFields.deletedAt }});
     }
 
-    if (options && options.deletedBy === true) {
-        schema.add({ deletedBy: Schema.Types.ObjectId });
+    if (options.deletedBy === true) {
+        schema.add({ deletedBy: { type: Schema.Types.ObjectId, index: indexFields.deletedBy }});
     }
 
     schema.pre('save', function (next) {
@@ -66,7 +98,7 @@ module.exports = function (schema, options) {
         next();
     });
 
-    if (options && options.overrideMethods) {
+    if (options.overrideMethods) {
         var overrideItems = options.overrideMethods;
         var overridableMethods = ['count', 'find', 'findOne', 'findOneAndUpdate', 'update'];
         var finalList = [];
@@ -136,7 +168,7 @@ module.exports = function (schema, options) {
             this.deletedBy = deletedBy;
         }
 
-        if (options && options.validateBeforeDelete === false) {
+        if (options.validateBeforeDelete === false) {
             return this.save({ validateBeforeSave: false }, callback);
         }